feat: simplify adding plugin
